+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](semver.md).
 ### Added
 * `times` extension for String
 * `TablePrinter` and table DSL to print Text Tables to any PrintStream
+* `prompt<T>` global function as Text Prompt with type parsing
 ### Changed
 * `isNumeric` extension on String uses start and end in regex `^\d+$`
 ### Deprecated